WL#12731: Improve database synchronization during mysqld startup [Part 1/6]
Added the column 'se_private_data' to the DD table Schemata.

The ndb version details has to be stored for every Schema in the DD so
that they can be compared later against the version information stored
in the ndb_schema table. For this purpose, we need a field in the DD's
Schemata table where the SE specific information can be stored. So, this
patch adds a field 'se_private_data' to the Schemata table. This field can
store version information as key value pairs. This patch also adds
auxiliary functions in the Schema class to read and write into the field.
These functions are also now implemented in the Schema_impl class.

Change-Id: I2f3a7aa0fdd85463db738ed649bf5b34fc3a4023