WL#12731: Improve database synchronization during mysqld startup [Part 2/6]
When a database is being created or altered, the server executes the
query and during that execution, calls the handler's binlog logger
through ha_binlog_log_query() method. The ndbcluster plugin currently
uses this callback function to distribute the DDL through schema
distribution. The version information of the database is generated at
this point and would be set to the Schema object in the DD. But in
mysql_create_db() method, this callback is called before the Schema
object is being created in the DD. Due to this, the ndbcluster plugin
is unable to set the version information to the DD object when the
callback is executed. This patch moves the ha_binlog_log_query()
function call after the creation of the Schema object in the DD.

Change-Id: I9fb5feb6d7b44032181f7dcbc920286d9663ea17