WL#12731: Improve database synchronization during mysqld startup [Part 4/6]
Added code to update Schema object's version from ndbcluster.

The Schema version can be updated by ndbcluster only during schema
distribution. In the distribution co-ordinator, Schema version is now
updated when the Server calls the engine's binlog logger function to
log the Schema operation. And in the distribution participant, the
Schema version is updated right after the DDL is executed in it. To
update the version, the new counter and node_id values are set to the
se_private_data of the Schema object in the relevant functions.

Change-Id: I126e2dd48516f385a558f99a95dc579cc2beabde