Bug#28971594: VIEW_CHANGE_LOG_EVENT MAY NOT BE LOGGED ON A OVERLOADED PRIMARY
On group changes, when a new member joins, all servers already in the
group try to log a View Change Log Event (VCLE) in the binary log for
the Group Replication recovery process.
This event however must be logged only when all local transactions
that preceded it are in the binary log and for this purpose there is a
wait in the plugin process.

If there was a timeout in this process, the event could be logged in a
improper place meaning the joiner would not receiving some of the
donor transactions.

RB: 21326