Bug#29031684 HANDLE_FATAL_SIGNAL (SIG=11) IN SQL_CONDITION::MESSAGE_TEXT Bug#28901919 INFORMATION_SCHEMA.TABLES COPIES TABLE_COMMENTS TO OTHER TABLES
Bug#29031684:
The problem was that we were reading diagnostic area when max_error_count
is set to zero. This lead to the problem reported.  This patch avoid
reading the diagnostic area and prepares the message string instead.

Bug#28901919:
The problem was that we were reading always the first sql condition in
diagnostic area, instead of last condition.  Due to this, I_S query picked
wrong error message whenever we found a invalid view. The problem is fixed
by preparing the error message string and not reading the DA, which works
even when max_error_count is set to zero. The test result of
main.information_schema_db is recorded with correct results now.

A test case is added in main.information_schema_cs.

Change-Id: I39deac6d4e3f2c07e81599e5677a5a9ffe370065