Bug#28955358 VIRTUAL LONGLONG FIELD_NEWDATE::VAL_DATE_TEMPORAL(): ASSERTION `!TABLE || (!TAB
The optimizer decides to do a filesort before reading OUTR.
So this filesort must store in its buffer all columns of OUTR needed
for its execution.
To determine the set of those columns, read_all_rows() does
first register_used_fields(), then
qep_tab->condition()->walk (this gathers the columns used by
the condition attached to the table, which condition is tested during
filesort).
As OUTR is materialized in a MEMORY-engine table, addon fields are not used
(see init_for_filesort()) (if they were used, then there would be no
problem).
register_used_fields() looks at addon fields, but as they are not used,
it doesn't help.
So it's really up to the walk() to catch the fields which the condition
uses. Here, the condition is:
WHERE ... IN (SELECT FROM (derived table referencing a column of OUTR) AS qn, ...
so the walk operates on IN, dives into the subquery, walks its SELECT list,
WHERE clause, etc, but misses the reference located inside 'qn'.
This can happen now that we have WL#461.

The same bug existed even before WL#461, with JSON_TABLE; a testcase is
added for it too.
It was also causing a problem in functional dependency detection; a
testcase is added for it.

Fix: Item_subselect::walk_body() must walk its FROM clause too. To this
aim, all its walk job is moved to functions SELECT_LEX_UNIT::walk
and SELECT_LEX::walk. For JSON_TABLE, Table_function_json::walk is
added.

An alternative was earlier suggested by Roy: when the walk on
qep_tab->condition() meets an Item_field, and this field is from a
materialized table, then walk the materialized query block.
But I think this won't always work: in the example added to the test
file:
SELECT 2 FROM (materialized derived table having outer ref) AS qn
the walk meets '2' and no column of 'qn'; but 'qn' will be
materialized, so its outer ref will be read, so it still needs some
read_set marking.
Moreover, it looks natural to me that, as outer-correlated derived
tables are some sort of subquery, so walk them just like we walk
Item_subselect.

Minor consequence of the patch: table_function.h needs access to
Item::enum_walk so has to include item.h. A second patch will declare
enum_walk in a better way.

Reviewed by: Evgeny Potemkin <evgeny.potemkin@oracle.com>