WL#529 Add CAST to FLOAT/DOUBLE/REAL
Aditional CAST function to support cast operations to FLOATING point
data types as specified by the SQL Standard. Also includes some
refactoring of the existing cast functions, in order to achieve a common
prefix for all the functions of this type(Item_typecast_*).
Item_func_binary was removed as it was unreachable.

Change-Id: I16bc88bbfaebfb639e2fda3deb37fb5c92470512