Bug#29291533 BASESTRING::SUBSTR() LACKS VALIDATION AND LEADS TO HEAP-OVERFLOW
BaseString::substr(), takes start and stop index as inputs and returns a new
string which will be a subset of the base string. BaseString::substr() does
not check if the length of the substring requested is within the length of the
base string. So, if an invalid inputs with substring's start or stop index
greater than the length of the basesting is requested, the function reads the
memory that does not belong to the process and leads to a heap overflow.

Example:

BaseString ff("ban");
ff.substr(0,8);
ff.substr(10,2);
ff.substr(8,16);

In above example size of base string is 3 ("ban"), but size of sub-string
requested is 8 which is invalid. In second case, the start index of the
sub-string is greater than the size of the base string. In third case, both
indexes are greater than the size of the base string.

Fixed by modifying the function to return a null string when the index of the
sub-string requested is not within the length of the base string.

Change-Id: I86a211dca129414e11de15366e6e02aa9e64f02d