Bug#29345751 - ASAN REPORTS MEMORY LEAKS IN SECONDARY ENGINE RELATED TO INNODB PARALLEL READ - 1
Issue
=====
The leak happens when there is an error during the offload after InnoDB has
created the partitions(sub-trees) required for the parallel read. There are
couple of issues related to this leak.

1) There are basically three stages to the offload operation - scan_start,
scan_run, scan_end. During scan_start we create the partitions (and hence the
`ranges` object for which we allocate memory) based on which contexts are
created in scan_run which the threads pick up for the parallel read, and in
scan_end we do the necessary cleanup.

The leak is seen when secondary engine aborts the offload operation after
scan_start (after finding out that the the table has already been offloaded, for
example). Although scan_end is called in this case the required cleanup is not
done - i.e., the `ranges` object in the created partitions are not freed.

2) Usually the cleanup related to these partitions is done in scan_run after
the parallel read, where we take the necessary steps to free the contexts that
have been created, based on the assumption that the ranges objects (for which
the memory has been created) present in the contexts and the partitions are the
same. Turns out this is not the case as the reference parameters are not
working as expected.

Fix
===
1) Do the necessary cleanup in scan_end where we free the partitions created as
part of scan_start. In the normal scenario the `ranges` object would be cleaned
up as we free the contexts in scan_run and in this specific case the cleanup of
partitions during scan_end would ensure that the memory related to it is freed
up.

2) Make the reference parameters work as expected so that the ranges partition
in both the partitions and the contexts are the same.

Reviewed-by: Sunny Bains <sunny.bains@oracle.com>

Change-Id: I523bc921ad0f7b3781aa6c52725e7b8a81638bfc