Bug#28590014 INNODB_DATA_FILE_PATH HAS A BAD DEFAULT VALUE
Issue
=====
Read-only sytem variables can now be persisted with SET PERSIST_ONLY clause.
But when the InnoDB system variables innobase_data_file_path and
innobase_temp_data_file_path are set to default using the same clause the
server fails to restart as the default value taken for these system variables
is NULL. This is mainly because the default value is set to NULL at the time
of system variable decalaration and we seem to update it to the actual default
value in function innodb_init_params().

Fix
===
Set the default value for system variables innobase_data_file_path and
innobase_temp_data_file_path during the declaration itself.

Reviewed-by: Debarun Bannerjee <debarun.banerjee@oracle.com> (Over IM)
Change-Id: I78e03371f26f370aa085d14dc8910ee77c1c8854