Bug #29240931 BACKWARD INCOMPATIBLE DESERIALIZATION OF MESSAGES FROM 5.7 ON 8.0.16
Problem
========================================================================
When a 5.7 sends a state exchange message to a 8.0.16 member, the last
8 bytes of the message are corrupted.

Analysis
========================================================================
WL #11610, released in 8.0.16, introduces message fragmentation into
GCS. When a member joins a group where fragmented messages are going
around, the joiner needs to recover the already delivered fragments
(before the member joined) so the joiner can reassemble the complete
message when the last fragment is delivered (after the member joins).
The fragment information is disseminated as part of the GCS state
exchange procedure, appended to the end of the state exchange message.

For backwards compatibility, WL #11610 also introduced the notion of the
group's communication protocol. Fragmentation is allowed when the
protocol is 8.0.16, and it is disallowed when the protocol is <8.0.16.
State exchange messages sent with protocol 8.0.16 contain the fragment
information, whereas state exchange messages sent with protocol <8.0.16
do *not* contain any fragment information.

When a member running MySQL 8.0.16 receives a state exchange message
from a member running MySQL <8.0.16, the state exchange message will
*not* contain any fragment information. However, the state exchange
message deserialization code was still assuming some fragment
information, and was consequently eating the last 8 bytes of the
exchanged data, which is the GR payload. Hence the inconsistent ways in
which the bug manifested itself, e.g. "random" values announced for the
lower_case_table_names property.

Additional details
------------------
The high-level layout of the state exchange message under protocol
8.0.16 is:

    +--------+------------------+----------------+
    | header | data to exchange | fragment info. |
    +--------+------------------+----------------+
                               /                  \
                              /                    \
      _______________________/                      \
     /                                               \
    +-------------+-...-+-------------+---------------+
    | fragment #1 |     | fragment #n | nr. fragments |
    +-------------+-...-+-------------+---------------+
                                       \______ ______/
                                              v
                                           8 bytes

The high-level layout of the state exchange message under protocol
<8.0.16 is:

    +--------+------------------+
    | header | data to exchange |
    +--------+------------------+

The implementation correctly identifies this message as not having
"fragment info.," but still thinks that "nr. fragments" is in the
message, so it chops up the last 8 bytes. This is because the
implementation calculates the "fragment info." size to be...

    nr. fragments * fragment size + 8 bytes

...independently of the protocol used. So it is considering the last 8
bytes to be "nr. fragments."
Unfortunately those last 8 bytes correspond in fact to the last 8 bytes
of "data to exchange," so we are eating the last 8 bytes of the GR
payload.

Solution
========================================================================
Calculate the "fragment info." size to be 0 if the protocol used is
<8.0.16.

Reviewed-by: Nuno Carvalho <nuno.carvalho@oracle.com>
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
RB: 21503