Bug#29232684: SLOW QUERY LOG DOESN'T LOG A SLOW CREATE INDEX WITH ADMIN STATEMENTS ENABLED
The query_time field in the slow log shows the overall time
the query took to execute (that is, the difference between the
wallclock times at the end and at the start of processing of the
statement).

The similarly named variable @@session.long_query_time currently
sets the minimum time a query must run after the locks are acquired
to count as a long query and potentially qualify for the slow log.

In unpatched trunk, the timestamp of when the locks were
acquired is updated each time locks are taken in mysql_lock_tables().

Due to the introduction of the data dictionary, scenarios where
locks are acquired more than once now need to be handled; if we
just let the server update the timestamp each time, the value
we'd compare to long_query_time would be the shortest of the
bunch; queries that the user would expect to be included into
the slow query log of their run-time might therefore not be
included.

Rather than just ignore locks acquired by the data dictionary,
we prevent the updating of utime_after_lock outright to cover
all scenarios. This fixes the immediate issue.

Approved by: Jens Even <jens.even.blomsoy@oracle.com>