Bug #29127873 NDB : AVOID TESTUPGRADE FAILURES DUE TO 7.6
The 7.6 data node is unable to restart from disk-based state recorded
by earlier versions.
This was an intentional limitation introduced to simplify the task
of modifying the disk-based LCP and Disk Data file formats.
However the upgrade testing in testUpgrade was not modified in
accordance, and so tests which expect a disk-based restart to work
all fail, making it harder to detect regressions against the
background noise of these failures.

This patch attempts to improve the situation :

  1) A new test called ShowVersion is introduced which will allow
     a quick visual check to see which actual versions were involved.
     This should help avoid a recurrance of the situation where
     testUpgrade is just testing various same-version restarts.

  2) A new mechanism is introduced to allow testUpgrade to determine
     which versions are involved in an up/downgrade at runtime prior
     to embarking on the upgrade.

  3) A new mechanism is introduced to allow NDBT tests to be SKIPPED,
     which is different to either PASSING or FAILING

  4) testUpgrade is modified to SKIP upgrade or downgrade tests which
     would fail due to the limitations of the 7.6 version
     (ATRT was extended to handle this SKIP result in
      WL#12865 SUPPORTS ATRT SKIPPED TESTS)

  5) testUpgrade is modified to fix a test added to check that 7.6
     refuses a disk-based restart - the test was ineffective.

This should result in failures related to the 7.6 data node
incompatibilities being removed from the results, allowing easier
identification of real regressions.

Non top-of-tree archived versions are used for some upgrade/downgrade
tests.  There can be failures with upgrades to/from those versions
until they are replaced with versions including this patch.

Reviewed-by : Tiago Alves <tiago.alves@oracle.com>