BUG#29111514 FLUSHING BINARY LOGS WITH GTID LEADS TO AN OOM ERROR
Description
-----------

When the `mysql.gtid_executed` table is opened read-only, rotating the binary
log will fail with an out-of-memory error.

Analysis
--------

During binlog rotation, prior to creating a new file, the GTID set contained in
the current binlog is meant to be dumped into the `mysql.gtid_executed` table.

If by some reason the access to the `mysql.gtid_executed` table fails, the
result of the failing function call is being always processed as an
out-of-memory error.

Fix
---

1) Properly return, from `Gtid_state::save_gtids_of_last_binlog_into_table`,
   different error codes for out-of-memory and denied access to the table.

2) Process `Gtid_state::save_gtids_of_last_binlog_into_table` return status and
   output different messages for the different return status.

3) Make `mysql.gtid_executed` accessible when `read_only` or `super_read_only`
   are ON.

4) If the binlog rotation is triggered by any other reason than
   `max_binlog_size` being reached and access to the GTID table is denied, keep
   logging to the current binlog and log a warning to the server log.

5) If the binlog rotation is triggered by `max_binlog_size` being reached and
   access to the GTID table is denied, follow rule for `binlog_error_action`.

ReviewBoard: 21338
Reviewed-by: Neha Kumari <neha.n.kumari@oracle.com>
Reviewed-by: Daogang Qu <bill.qu@oracle.com>