Bug #20344149 NDB : MAXBUFFEREDEPOCHS BACKOFF / HEADROOM
The MaxBufferedEpochs config parameter is used as a protection
on the data nodes to avoid excessive buffering of row changes
due to lagging NdbApi event Api subscribers.

When one or more subscriber's epoch acknowledgements lag by
this number of epochs, an asynchronous disconnection is triggered
which eventually allows the data node to release the buffer
space used for the subscriber's subscription(s).

As disconnection is asynchronous, it may be the case that it has
not completed before further new epochs are completed on the
data node.  This can result in new epochs not being able to
seize a GCP completion record, generating warnings like :

[ndbd] ERROR    -- c_gcp_list.seize() failed...

and subsequently

[ndbd] WARNING  -- ACK wo/ gcp record...

This fix :
 - Modifies the sizing of the GCP completion record pool to ensure
   that there is always some extra 'headroom' to account for the
   asynchronous nature of the disconnect processing without getting
   c_gcp_list seize failures
 - Modifies the wording of the MaxBufferedEpochs warning to avoid
   the contradictory phrase "100>100"

Manually tested.

Reviewed-by: Martin Skold <martin.skold@oracle.com>