Bug #29296645 PERFORMANCE_SCHEMA.DATA_LOCKS.LOCK_DATA SHOULD SHOW PK COLUMNS FOR UNIQUE SEC
For RECORD locks, the performance_schema.data_locks.lock_data column displays comma-separated
list of columns which identify the record being locked.
For PRIMARY key and for non-UNIQUE indices the information shown in this column was enough to
identify the record uniquely.
In particular for non-UNIQUE secondary index the columns shown included clustered index columns.
However for UNIQUE secondary index, we have only shown the columns on which the secondary index
was explictilty defined, hoping that the uniqueness of the index is enough to guaranttee that
there is only one record matching such entry.

InnoDB uses multi-version concurrency control, which means that there might be multiple versions
of the row with the same primary key, some of them deleted, at the same time.
If there is a secondary index defined, then it will contain entries for some recent versions of
the row to facilitate fast access to them in MVCC.
This in turn may lead to situation where a UNIQUE secondary index contains multiple entries with
the same values of unique columns, which is completely fine, as long as they differ by clustered
index values, that is, they point to different rows in clustered index.
The clustered index columns are therefore a useful piece of information, as they provide the "link"
between secondary index and clustered index.

This patch makes the clustered index columns visibile in lock_data column also for UNIQUE index.

RB:21532

Reviewed-by: Marc Alff <marc.alff@oracle.com>