Bug#29189532: FLUSH OF BUFFERED MESSAGES TO ERROR LOG SEEMS TO BE WITHHELD DURING DD UPGRADE
During start-up, we buffer log-info until a) we have
basic info for the built-in logger (what file to log
to, verbosity, and so on), and b) advanced info (any
logging components to load, any configuration for
them, etc.).

As a failsafe, if start-up takes very long and a time-out
is reached before reaching b) and we actually have something
worth reporting (e.g. errors, as opposed to info), we try to
keep the user informed by using the basic logger configured
in a), while going on buffering all info and flushing it to
any advanced loggers when b) is reached.

The buffered logger previously did two things:

1) Checks and, if needed, update the time-out, and call
   the flush functions as needed.

2) Save log-event for later filtering and output to
   log sinks that are loadable components.

Patch adds a new function log_sink_buffer_check_timeout()
that gives other subsystems access to functionality 1)
without having to submit log information.

Functions that have the potential to run for a very long
time during start-up, such as server upgrade, can now
check whether the buffered logging time-out has been
exceeded, in which case the log information will be
flushed as described.

The function should not be called once the component
subsystem is fully initialized and any external components
are loaded. (As a rule of thumb, if your code runs after
InnoDB initializing and data dictionary upgrade, don't use
this.)

For subsystems that run rarely (i.e. not at every start-up),
it may be worthwhile to shun this function and issue a
SYSTEM severity message of the form "%d tables processed"
every x (e.g. 64 or so) tables (or even "%d/%d tables
processed" if we know the total in advance).

Approved by: Sivert Sorumgaard <sivert.sorumgaard@oracle.com>