Bug#29365552: LEFTOVER SYSTEM TABLES AFTER 5.7 => 8.0 IN-PLACE UPGRADE
The upgrade from 5.7 to 8.0 runs on a Bootstrap thread and so, innodb post
DDL hooks don't run at this stage. Due to this, even if a DROP TABLE statement
is executed, the physical tablespace file is not deleted.

This patch adds code to manually delete the tablespace files of
innodb_table_stats_backup and innodb_index_stats_backup tables at the
end of a successful upgrade.

Change-Id: Ia81e2ccf80fc425418d9e06bddace6b0e6402ccd