Bug#29356132: OPTIMIZED-AWAY SUBQUERY IN PREPARED STATEMENT               CAUSES ASSERT FAILURE IN EXECUTION
Problem:
While flattening a subquery, if an always false predicate is present,
optimizer does not perform any kind of transformation
(semi-join or otherwise).This results in a subquery that is not prepared
(Asserts later when executed).

Solution:
When an always false predicate is present for a subquery, unlink this
subquery's query expression from the query block.

Reviewed by: Roy Lyseng <roy.lyseng@oracle.com>