Bug#29110719 FAILURE OF NDB.NDB_DBUG_LOCK TEST
Removes the test ndb_dbug_lock.

ndb_dbug_lock.test is a simple test which tries to verify that a table cannot
be altered when it is locked. The lock order from the time the test was
disabled has been changed(Currently GSL lock is taken before any kind of
locks). So, the testcase fails with "Test case timeout after 900 seconds",
because the test tries to serialize the alter commands.

The tests for how global schema lock is combined with LOCK TABLES is
implemented in ndb_global_schema_lock.test. So, ndb_dbug_lock.test is declared
obsolete and is removed.

Change-Id: I9bb8498cae1a8feea9b3fd286c23a91f8e6b37e0