BUG#29210040 [1/2] WINDOWS: ROUTER FAILS TO START AFTER BUILDING AND INSTALLING FROM SOURCES
This patch fixes an issue on Windows, where Router may fail to start due
to not being able to load plugins.

This happened because CMake files pointed Router to look for plugins in
one directory (ROUTER_INSTALL_PLUGINDIR), and then at runtime (C++ code)
this path was augmented, depending on whether this path was different
from a hard-coded path. Not only was this a hack, but it also had a bug:
this string comparison was case sensitive, while the paths on Windows
are not, and thus in some cases it returned true while others false, for
the exact same path being compared. Building and installing from
sources, if done in a certain way, just happened to trigger this bug.

Fixed by removing faulty logic and setting ROUTER_INSTALL_PLUGINDIR to
correct value at the CMake level.

RB: 21696