Bug#29389168 DISK DATA TABLE UNUSABLE ON PARTICIPANT MYSQLD AFTER ALTER TABLESPACE
Problem:
--------

After an ALTER TABLESPACE query was distributed, both DML and DDL queries
on a previously created table using the aforementioned tablespace failed
with the following error:

ERROR 3508 (HY000): Dictionary object id (10) does not exist.

The dictionary object id specified refers to the tablespace. Schema
distribution of ALTER TABLESPACE involves dropping the old object from DD
in a participant MySQL Server and creating a new one (with a different
dictionary object id). However, the table object in DD of the participant
MySQL server still referred to the old tablespace id which rendered it
unusable on the participants.

Fix:
----

Retrieve and store the tables in the tablespace prior to the creation
of the new tablespace. Update these tables with the new object id
of the tablespace after the new tablespace has been created in the DD

Change-Id: I35d63f8ab3275ebf320ca39ff0b48119b39bf7fa