Bug#29465794 ASAN failure in ndb_restore while calling substr.
Problem
=======

In ndb_restore a 8 character long substring of table name is extracted
which then is compared to NDB$BLOB to see if the table is a blob part
table or not.

BaseString::substr() tried to access eigth characters in source string even
if that string was shorter, this read beyond buffer was detected by ASAN.

Solution
========

Implement BaseString::starts_with() and use that to see if prefix NDB$BLOB
matches table name.

Change-Id: I0b2be72973ccb30ecbc1e7ebd261aa86bd66aae1