Bug #29545644: ALLOCATE ITERATORS ON THE MEM_ROOT
Make all allocators consistently be allocated on MEM_ROOT. We add
a new wrapper called NewIterator() to encapsulate this pattern
(in WL #4168, it will be used to construct special timing
iterators for EXPLAIN ANALYZE), and use it everywhere.

In particular, this meant that READ_RECORD is now gone. It went from
being an ad-hoc union in the pre-iterator days, to a real union of
all possible basic RowIterators earlier (with some extra space thrown
in for SortingIterator), into just unique_ptr to a RowIterator now.
This should make us allocate less memory for the QEP_TAB in the common
case, as well as reducing complexity a bit, at the cost of having to
allocate the iterators on a MEM_ROOT (which is usually very cheap).

setup_read_record(), init_read_record() and setup_read_record_idx()
are replaced with variants that return an iterator directly, instead
of taking in an output READ_RECORD pointer.

The union pattern persists in SortingIterator for the time being,
since these never need to be wrapped in TimingIterators for WL #4168
(they're not visible in EXPLAIN).

Change-Id: I15b0dc9df26ffda66c4aaed07dcc5482aac2e7ff