Bug#28942965: MYSQL EXITS WITH CHECK_DATE HAS AN INVALID               POINTER
Description: Server exits for derived table queries with
ORDER BY clause.

Analysis:- The problem may occur when a derived table with
an ORDER BY clause is merged into an outer query, when the
columns from ORDER BY are not also referenced in outer
query. The problem is due to missing to mark the column
underlying 'field2' in the read set of table t1.

'field2' is used in ORDER BY inside a derived table.
'field2' references a column from the SELECT list. But since
fields from SELECT list in a derived table are not marked in
the read_set when they are first resolved, we need to make
sure those fields are marked when used for other purposes,
such as in ORDER BY. Notice that we only want to mark
columns when merging into an outermost query expression,
and not into another derived table.

Fix:- Back porting the fix for Bug#26542829 from mysql-8.0
to mysql-5.7.

Change-Id: I3805c7017a9c8cfbb003265b9ff54bf10f33305b