Bug #28971954 MOVE GCS_XCOM_CONTROL_INTERFACE-T.CC TO THE UNIT TESTS DIRECTORY
Problem and analysis
========================================================================
The gcs_xcom_control_interface-t.cc test file is in the smoke tests'
directory but it should be in the unit tests'.
The file's tests take around ~300 seconds, which is too much.
The reason is that we wait for the suspicions processing thread to wake
up on its own, and terminate, when we we leave the group.

Solution
========================================================================
Proactively wake up the suspicions processing thread when we leave the
group, instead of waiting for the thread to wake up on its own. This
cuts down the duration of several tests substantially.

Move the gcs_xcom_control_interface-t.cc file to the unit tests'
directory and merge it with the recently created
gcs_xcom_control_interface_unit-t.cc.

Suppress a message about misconfigured local_address on
gr_autorejoin_stop_gr. This message is expected because the test
misconfigures local_address to trigger the rejoin procedure.
The hypothesis is that the message did not appear before due to the
artificial delay in the leave operation. (Waiting for the suspicions
processing thread to wake up.)

Suppress a message about view change timeout on gr_recovery_no_donors.
This message appears because there is a circular dependency between the
delivery of view change notifications and an ongoing recovery procedure.
Specifically, the view change notifications are only sent after the
recovery procedure terminates, but the recovery procedure only
terminates after receiving the view change notification: both wait for
each other until they time out.
The message did not appear before due to the artificial delay in the
leave operation. (Waiting for the suspicions processing thread to wake
up.) The recovery procedure would finish before the view change, whereas
now the view change is delivered while the recovery procedur is ongoing.
A bug report was opened to deal with this situation, whose patch should
remove the suppressions.

Reviewed-by: Andre Negrao <andre.negrao@oracle.com>
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
RB: 21373