Bug #29297652: STACK OVERFLOW IN READ_OK_EX() WITH SESSION_STATE_CHANGED
RB #21624

A static buffer was used to receive the charset name without checking
if its size suffices.

Fixed by adjusting the size of the buffer so it can take the longest
possible character set name and then copying just the part of the
incoming charset name that fits into the buffer.