BUG#29411832 REWRITE LOG_BLOCK_FIRST_REC_GROUP DURING RECOVERY MAY BE DANGEROUS.
During the recovery we start the redo log, parse log records and flush dirty pages.
However, just before the last round of flushing of the dirty pages, we start redo
threads and enable ibuf merges to happen. These merges may generate new redo records,
even though the recovery is still not ended then. The redo threads would write the
new redo records to disk.

If we crashed during the last round of recovery from crash, after such new redo records
were written to disk, but before we flushed the recovered pages, we could have a data loss.

This data loss would only happen if the checkpoint_lsn was within the last block of the
redo (the same block as current_lsn). The reason of the possible data loss is that we
re-initialize the first_rec_group field in the last block, when we start the redo log,
and such re-initialized value could be rewritten by the redo threads before recovery
was ended. The same field is used when recovery starts to find the beginning of the
recovery, starting from the block which contains checkpoint_lsn.

The fix is not to re-initialize the first_rec_group unless it seems to be corrupted
after the recovery. To determine that, we track expected values of first_rec_group
during recovery, whenever we progress with recovered_lsn.

Note that in all versions of MySQL earlier than 8.0.5, the bug could manifest in
a different way. We were never re-initializing the first_rec_group after recovery
ended with success. If the last block had a corrupted first_rec_group on disk,
recovery could still end with success if it was starting in earlier blocks.
After recovery ended, we would proceed with normal workload, inserting records
to such a corrupted block (with invalid first_rec_group). If we crashed, data loss
was possible. All because of the corrupted redo log, which was corrupted before we
started the MySQL, which could be misleading as the recovery stated that all was
fine within the redo files.

Special thanks to Kang Wang who has found and reported the bug.

RB: 21619