WL #4168: Implement EXPLAIN ANALYZE [patch 1/2, costs on EXPLAIN FORMAT=TREE]
Apply existing costs and row estimated from the planner to the iterators,
giving EXPLAIN FORMAT=tree the same cost information as is already available in
the JSON output. Only certain iterators have this information (e.g. sorting
does not, and we don't have full control over materialization costs), so many
will be left blank until we have an optimizer that works natively with
iterators.

For the time being, we store the information directly on the iterators.
This should also change when we rewrite the optimizer.

Change-Id: Iedb0571bc3c97e645c8052e5bf26f71f3029a65a